packages/extractor/requirements.txt
python-pptx
pika
boto3
dotenv
